try to work around non-API nonsense
