+++ com.ibm.wala.cast.js.rhino/src/test/java/com/ibm/wala/cast/js/rhino/test/HTMLCGBuilder.java
@@ -25,6 +25,7 @@ import com.ibm.wala.ipa.callgraph.CallGraph;
 import com.ibm.wala.ipa.callgraph.CallGraphBuilderCancelException;
 import com.ibm.wala.ipa.callgraph.propagation.InstanceKey;
 import com.ibm.wala.ipa.callgraph.propagation.PointerAnalysis;
+import com.ibm.wala.util.MonitorUtil.IProgressMonitor;
 import com.ibm.wala.util.NullProgressMonitor;
 import com.ibm.wala.util.io.CommandLine;
 import com.ibm.wala.util.io.FileUtil;
@@ -88,7 +89,8 @@ public class HTMLCGBuilder {
       // the code below belongs somewhere else!!!
       // the bound of 4 is what is needed to pass our current framework tests
       builder.setContextSelector(new RecursionCheckContextSelector(builder.getContextSelector()));
-      ProgressMaster master = ProgressMaster.make(new NullProgressMonitor(), timeout * 1000, false);
+      IProgressMonitor master =
+          ProgressMaster.make(new NullProgressMonitor(), timeout * 1000, false);
       if (timeout > 0) {
         master.beginTask("runSolver", 1);
       }
